+++ spec/lib/converter_spec.rb
@@ -3,12 +3,27 @@ require "spec_helper"
 module MinitestToRspec
   RSpec.describe Converter do
 
+    # Fixtures that represent rails use cases.  These will
+    # instantiate Converter with `rails: true`.
+    RAILS_FIXTURES = [15]
+
+    # The path to `spec/fixtures`
     SPEC_FIXTURES = File.join(__dir__, '..', 'fixtures')
+
+    # Directories under `spec/fixtures`
     FIXTURE_DIRS = Dir.glob("#{SPEC_FIXTURES}/*")
 
     def convert(input, options = nil)
-      options ||= { rails_helper: false }
+      options ||= { rails: false }
       described_class.new(options).convert(input)
+    end
+
+    def fixture_number(fixture)
+      File.basename(fixture).split('_').first.to_i
+    end
+
+    def rails?(fixture)
+      RAILS_FIXTURES.include?(fixture_number(fixture))
     end
 
     def read_input(fixture)
@@ -23,14 +38,15 @@ module MinitestToRspec
       FIXTURE_DIRS.each do |fixture|
         it "converts: #{fixture}" do
           expected = read_output(fixture).strip
-          calculated = convert(read_input(fixture)).strip
+          options = { rails: rails?(fixture) }
+          calculated = convert(read_input(fixture), options).strip
           expect(calculated).to eq(expected)
         end
       end
 
-      it "supports rails_helper option" do
+      it "supports rails option" do
         expect(
-          convert("require 'test_helper'", rails_helper: true)
+          convert("require 'test_helper'", rails: true)
         ).to eq('require("rails_helper")')
       end
     end