+++ spec/lib/subprocessors/call_spec.rb
@@ -15,21 +15,23 @@ module MinitestToRspec
           RubyParser.new.parse(ruby)
         end
 
-        def process(input, rails_helper = false)
-          described_class.process(input, rails_helper)
+        def process(input, rails = false)
+          described_class.process(input, rails)
         end
 
-        context "rails_helper is false" do
-          it "replaces test_helper with spec_helper" do
-            input = exp(:require, "test_helper")
-            expect(process(input, false)).to eq(exp(:require, "spec_helper"))
+        describe "rails option" do
+          context "option is false" do
+            it "replaces test_helper with spec_helper" do
+              input = exp(:require, "test_helper")
+              expect(process(input, false)).to eq(exp(:require, "spec_helper"))
+            end
           end
-        end
 
-        context "rails_helper is true" do
-          it "replaces test_helper with rails_helper" do
-            input = exp(:require, "test_helper")
-            expect(process(input, true)).to eq(exp(:require, "rails_helper"))
+          context "option is true" do
+            it "replaces test_helper with rails_helper" do
+              input = exp(:require, "test_helper")
+              expect(process(input, true)).to eq(exp(:require, "rails_helper"))
+            end
           end
         end